Fixed resource pass trough logic for players.
Unit dropp off resources tried to send resources to building reference with wrong interface function.
Renamed the function from UnitDroppingOfResources to DropOffResourcesToBuilding to clarify the action.
OwningPlayer initialization was not working in buildingparentclass.
This meant the increment player resources message was not reseived by the player controller.
Added the player reference to setup player camp macro in player cotroller for the initally spawned actors.