Completed issue #5 and #26, new units, refactoring, bug fixes
Refactored the player controller unit commands.
Modified resource interface to have UnitInteractedWith (unit interface has same function) so that resource only needs to implement resource interface. Fixed a wrong state change in resource unit reached. Added a basic body for parent unit. Deleted the one in child actors. Added FilterUnitsToCommand and FilterAllUnitsToCommand function to UnitHelperFunctions. Changed the SelectedUnits in marqueehud to be of type UnitParentClass. Refactored MarqueeHud to filter out only unit types for selected units array. Fixed a bug in unit if it tries to reach a null resource or tried to interact with it for first time but it was already destroyed.