Fixed a bug in UnitState. Fixed a bug in SpawnUnit
Units state is now set as a casted reference in player controller. Deleted the interface function use on the units state and made the increment, decrement and get resources member functions. They now allow automatic from server to client calls, multicast and to server call. Spawn unit now holds on to active unit in production to hold to the units spawning if the unit count is max for the player.